projects/01_rag_baseline/requirements.txt
langchain
langchain-community
langchain-openai
langchain-text-splitters
chromadb
python-dotenv
pytest
